# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authentication System API",
    "version": "1.0.0",
    "description": "API documentation for the Authentication System with Admin Roles"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/subscription/{id}": {
      "get": {
        "summary": "Get a subscription by ID",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Subscription ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription data"
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      },
      "put": {
        "summary": "Update a subscription (Admin only)",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Subscription ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string"
                  },
                  "planId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription updated"
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      },
      "delete": {
        "summary": "Cancel (delete) a subscription (Admin only)",
        "tags": [
          "Subscriptions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Subscription ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription cancelled"
          },
          "404": {
            "description": "Subscription not found"
          }
        }
      }
    },
...
    "/api/subscription/plans/{id}": {
      "get": {
        "summary": "Get a subscription plan by ID",
        "tags": [
          "SubscriptionPlans"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription plan details"
          },
          "404": {
            "description": "Plan not found"
          }
        }
      }
    },
    "/api/subscription/me": {
      "get": {
        "summary": "Get current user's subscription",
        "tags": [
          "SubscriptionPlans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User's subscription info"
          }
        }
      }
    },
    "/api/subscription/create": {
      "post": {
        "summary": "Create a subscription",
        "tags": [
          "SubscriptionPlans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "planId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscription created"
          }
        }
      }
    },
    "/api/subscription/can-download/{assetId}": {
      "get": {
        "summary": "Check if a user can download a specific asset",
        "tags": [
          "SubscriptionPlans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "assetId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download permission result"
          }
        }
      }
    },
    "/api/subscription/record-download": {
      "post": {
        "summary": "Record a download attempt",
        "tags": [
          "SubscriptionPlans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "assetId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Download recorded"
          }
        }
      }
    },
    "/api/subscription/create-plans": {
      "post": {
        "summary": "Create a new subscription plan",
        "tags": [
          "SubscriptionPlans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "duration": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan created"
          }
        }
      }
    },
    "/api/subscription/update-plans/{id}": {
      "put": {
        "summary": "Update an existing subscription plan",
        "tags": [
          "SubscriptionPlans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "duration": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan updated"
          }
        }
      }
    },
...
    "/api/subscription/admin/all": {
      "get": {
        "summary": "Get all user subscriptions (admin only)",
        "tags": [
          "SubscriptionPlans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all subscriptions"
          }
        }
      }
    },
    "/api/subscription/admin/user/{userId}": {
      "get": {
        "summary": "Get subscription history of a user",
        "tags": [
          "SubscriptionPlans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User subscription history"
          }
        }
      }
    },
    "/api/subscription/admin/{subscriptionId}/status": {
      "patch": {
        "summary": "Update the status of a subscription",
        "tags": [
          "SubscriptionPlans"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "subscriptionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "cancelled",
                      "expired"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription status updated"
          }
        }
      }
    },
...
  },
...
}